perf: improved the ranking commands
