+++ src/main/java/me/roinujnosde/titansbattle/managers/CommandManager.java
@@ -6,6 +6,7 @@ import co.aikar.commands.PaperCommandManager;
 import me.roinujnosde.titansbattle.TitansBattle;
 import me.roinujnosde.titansbattle.commands.ChallengeCommand;
 import me.roinujnosde.titansbattle.commands.ConfigCommands;
+import me.roinujnosde.titansbattle.commands.RankingCommand;
 import me.roinujnosde.titansbattle.commands.TBCommands;
 import me.roinujnosde.titansbattle.commands.completions.*;
 import me.roinujnosde.titansbattle.commands.conditions.*;
@@ -105,6 +106,7 @@ public class CommandManager extends PaperCommandManager {
         registerCommand(new TBCommands());
         registerCommand(new ChallengeCommand());
         registerCommand(new ConfigCommands());
+        registerCommand(new RankingCommand());
     }
 
     private <T> void registerContext(AbstractIssuerOnlyContextResolver<T> resolver) {
